requirements.txt
setuptools
mysql-connector-python
rich-click
rich
questionary